pyproject.toml
[tool.poetry]
name = "run_tracker"
version = "0.1.0"
description = ""
authors = ["Ben Dixon <uk.bendixon@gmail.com>"]
package-mode = false


[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
garmindb = "^3.6.3"
notebook = "^7.3.2"
snakemd = "^2.3.0"
openai = "^1.66.3"
numpy = "^2.2.3"
pytest = "^8.3.5"
pre-commit = "^4.1.0"
asyncio = "^3.4.3"
pytest-mock = "^3.14.0"
flask = "^3.1.0"
markdown = "^3.7"
isort = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "src"
